Move auth selectors to their own file; remove old references to use in test
