+++ src/containers/PrivateRoute.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 
-import { selectGetCurrentUserIsLoading, selectIsLoggedIn } from 'store/auth/reducer';
+import { selectGetCurrentUserIsLoading, selectIsLoggedIn } from 'store/auth/selectors';
 import { selectLoggedInUser } from 'store/entities/selectors';
 import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import getRoleTypesFromRoles from 'utils/user';
