+++ src/scenes/Legalese/index.jsx
@@ -7,7 +7,6 @@ import PropTypes from 'prop-types';
 import { getFormValues } from 'redux-form';
 
 import { reduxifyWizardForm } from 'shared/WizardPage/Form';
-import { selectGetCurrentUserIsSuccess } from 'store/auth/reducer';
 import CertificationText from './CertificationText';
 import Alert from 'shared/Alert';
 import { SwaggerField } from 'shared/JsonSchemaForm/JsonSchemaField';
@@ -21,6 +20,7 @@ import { completeCertificationText } from './legaleseText';
 import SectionWrapper from 'components/Customer/SectionWrapper';
 import { setFlashMessage } from 'store/flash/actions';
 import { selectCurrentPPM } from 'store/entities/selectors';
+import { selectGetCurrentUserIsSuccess } from '../../store/auth/selectors';
 
 const formName = 'signature-form';
 const SignatureWizardForm = reduxifyWizardForm(formName);
